+++ packages/gatsby/cache-dir/__tests__/minimal-config.js
@@ -9,6 +9,8 @@ it(
       path.join(__dirname, `..`),
       `--config-file`,
       path.join(__dirname, `.babelrc`),
+      `--ignore`,
+      `**/__tests__`,
     ]
 
     const spawn = child.spawn(process.execPath, args)