+++ packages/gatsby/src/utils/websocket-manager.js
@@ -92,6 +92,7 @@ const getRoomNameFromPath = (path: string): string => `path-${path}`
 class WebsocketManager {
   pageResults: QueryResultsMap
   staticQueryResults: QueryResultsMap
+  errors: Map<string, QueryResult>
   isInitialised: boolean
   activePaths: Set<string>
   programDir: string
@@ -101,6 +102,7 @@ class WebsocketManager {
     this.activePaths = new Set()
     this.pageResults = new Map()
     this.staticQueryResults = new Map()
+    this.errors = new Map()
     this.websocket
     this.programDir
 
@@ -108,6 +110,7 @@ class WebsocketManager {
     this.getSocket = this.getSocket.bind(this)
     this.emitPageData = this.emitPageData.bind(this)
     this.emitStaticQueryData = this.emitStaticQueryData.bind(this)
+    this.emitError = this.emitError.bind(this)
   }
 
   init({ server, directory }) {
@@ -131,6 +134,15 @@ class WebsocketManager {
         this.websocket.send({
           type: `staticQueryResult`,
           payload: result,
+        })
+      })
+      this.errors.forEach((message, errorID) => {
+        this.websocket.send({
+          type: `overlayError`,
+          payload: {
+            id: errorID,
+            message,
+          },
         })
       })
 
@@ -194,10 +206,21 @@ class WebsocketManager {
   }
 
   emitPageData(data: QueryResult) {
+    this.pageResults.set(data.id, data)
     if (this.isInitialised) {
       this.websocket.send({ type: `pageQueryResult`, payload: data })
     }
-    this.pageResults.set(data.id, data)
+  }
+  emitError(id: string, message?: string) {
+    if (message) {
+      this.errors.set(id, message)
+    } else {
+      this.errors.delete(id)
+    }
+
+    if (this.isInitialised) {
+      this.websocket.send({ type: `overlayError`, payload: { id, message } })
+    }
   }
 }