+++ www/src/components/ecosystem/ecosystem-board.js
@@ -1,10 +1,14 @@
-import React from "react"
+import React, { Component } from "react"
 import PropTypes from "prop-types"
 import styled from "react-emotion"
 
 import EcosystemSection from "./ecosystem-section"
 
 import presets from "../../utils/presets"
+import {
+  setupScrollersObserver,
+  unobserveScrollers,
+} from "../../utils/scrollers-observer"
 
 const EcosystemBoardRoot = styled(`div`)`
   display: flex;
@@ -20,46 +24,60 @@ const EcosystemBoardRoot = styled(`div`)`
   }
 `
 
-const EcosystemBoard = ({
-  icons: { plugins: PluginsIcon, starters: StartersIcon },
-  starters,
-  plugins,
-}) => (
-  <EcosystemBoardRoot>
-    <EcosystemSection
-      title="Plugins"
-      description="Plugins are packages that extend Gatsby sites. They can source content, transform data, and more!"
-      subTitle="Featured Plugins"
-      icon={PluginsIcon}
-      links={[
-        { label: `Browse Plugins`, to: `/plugins/` },
-        {
-          label: `Plugin Authoring`,
-          to: `/docs/plugin-authoring/`,
-          secondary: true,
-        },
-        { label: `Plugin Docs`, to: `/docs/plugins/`, secondary: true },
-      ]}
-      featuredItems={plugins}
-    />
-    <EcosystemSection
-      title="Starters"
-      description="Starters are Gatsby sites that are preconfigured for different use cases to give you a head start for your project."
-      subTitle="Featured Starters"
-      icon={StartersIcon}
-      links={[
-        { label: `Browse Starters`, to: `/starters/` },
-        { label: `Starter Docs`, to: `/docs/starters/`, secondary: true },
-      ]}
-      featuredItems={starters}
-    />
-    <EcosystemSection
-      title="External Resources"
-      description="A curated list of interesting Gatsby community projects and learning resources like podcasts and tutorials."
-      links={[{ label: `Browse Resources`, to: `/docs/awesome-gatsby/` }]}
-    />
-  </EcosystemBoardRoot>
-)
+class EcosystemBoard extends Component {
+  componentDidMount() {
+    setupScrollersObserver()
+  }
+
+  componentWillUnmount() {
+    unobserveScrollers()
+  }
+
+  render() {
+    const {
+      icons: { plugins: PluginsIcon, starters: StartersIcon },
+      starters,
+      plugins,
+    } = this.props
+
+    return (
+      <EcosystemBoardRoot>
+        <EcosystemSection
+          title="Plugins"
+          description="Plugins are packages that extend Gatsby sites. They can source content, transform data, and more!"
+          subTitle="Featured Plugins"
+          icon={PluginsIcon}
+          links={[
+            { label: `Browse Plugins`, to: `/plugins/` },
+            {
+              label: `Creating Plugins`,
+              to: `/docs/plugin-authoring/`,
+              secondary: true,
+            },
+            { label: `Using Plugins`, to: `/docs/plugins/`, secondary: true },
+          ]}
+          featuredItems={plugins}
+        />
+        <EcosystemSection
+          title="Starters"
+          description="Starters are Gatsby sites that are preconfigured for different use cases to give you a head start for your project."
+          subTitle="Featured Starters"
+          icon={StartersIcon}
+          links={[
+            { label: `Browse Starters`, to: `/starters/` },
+            { label: `Using Starters`, to: `/docs/starters/`, secondary: true },
+          ]}
+          featuredItems={starters}
+        />
+        <EcosystemSection
+          title="External Resources"
+          description="A curated list of interesting Gatsby community projects and learning resources like podcasts and tutorials."
+          links={[{ label: `Browse Resources`, to: `/docs/awesome-gatsby/` }]}
+        />
+      </EcosystemBoardRoot>
+    )
+  }
+}
 
 EcosystemBoard.propTypes = {
   icons: PropTypes.object,