+++ www/src/views/starter-library/index.js
@@ -1,7 +1,9 @@
 import React, { Component } from "react"
-import Helmet from "react-helmet"
+import { Helmet } from "react-helmet"
 import Layout from "../../components/layout"
+import Unbird from "../../components/unbird"
 import RRSM from "../../utils/reach-router-state-manager"
+import queryString from "query-string"
 
 import FilteredStarters from "./filtered-starters"
 
@@ -11,13 +13,16 @@ class StarterLibraryPage extends Component {
     return JSON.stringify(this.props) !== JSON.stringify(nextProps)
   }
   render() {
-    const { location, urlState } = this.props
+    const { location } = this.props
+    const urlState = queryString.parse(location.search)
+
     const filtersApplied =
-      urlState.s !== ``
+      urlState.s !== undefined
         ? urlState.s // if theres a search term
         : urlState.d && !Array.isArray(urlState.d)
           ? urlState.d // if theres a single dependency
-          : `Showcase` // if no search term or single dependency
+          : `Library` // if no search term or single dependency
+
     return (
       <Layout location={location}>
         <Helmet>
@@ -39,12 +44,26 @@ class StarterLibraryPage extends Component {
           <meta name="twitter.label1" content="Reading time" />
           <meta name="twitter:data1" content={`1 min read`} />
         </Helmet>
-        <FilteredStarters {...this.props} />
+        <RRSM
+          {...this.props}
+          location={location}
+          render={({ setURLState, urlState }) => (
+            <FilteredStarters
+              {...this.props}
+              setURLState={setURLState}
+              urlState={urlState}
+            />
+          )}
+          defaultSearchState={{ v: [`2`] }}
+        />
+        <Unbird
+          dataSetId="5c113a828240aa564734d954"
+          publicKey={process.env.UNBIRD_FEEDBACK_KEY_STARTERLIB}
+          feedbackPrompt="Have feedback on the Starter Library?"
+        />
       </Layout>
     )
   }
 }
 
-export default RRSM({ s: ``, c: [], d: [], v: [], sort: `recent` })(
-  StarterLibraryPage
-)
+export default StarterLibraryPage