+++ packages/gatsby-plugin-layout/CHANGELOG.md
@@ -3,6 +3,32 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.0.11"></a>
+
+## [1.0.11](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-layout@1.0.10...gatsby-plugin-layout@1.0.11) (2018-12-21)
+
+**Note:** Version bump only for package gatsby-plugin-layout
+
+<a name="1.0.10"></a>
+
+## [1.0.10](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-layout@1.0.9...gatsby-plugin-layout@1.0.10) (2018-11-29)
+
+**Note:** Version bump only for package gatsby-plugin-layout
+
+<a name="1.0.9"></a>
+
+## [1.0.9](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-layout@1.0.8...gatsby-plugin-layout@1.0.9) (2018-11-14)
+
+### Bug Fixes
+
+- **docs:** update broken links with working links ([#9912](https://github.com/gatsbyjs/gatsby/issues/9912)) ([e9f2a6f](https://github.com/gatsbyjs/gatsby/commit/e9f2a6f))
+
+<a name="1.0.8"></a>
+
+## [1.0.8](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-layout@1.0.7...gatsby-plugin-layout@1.0.8) (2018-11-08)
+
+**Note:** Version bump only for package gatsby-plugin-layout
+
 <a name="1.0.7"></a>
 
 ## [1.0.7](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-layout@1.0.6...gatsby-plugin-layout@1.0.7) (2018-11-05)