+++ packages/gatsby-plugin-page-creator/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="2.0.5"></a>
+
+## [2.0.5](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-page-creator@2.0.4...gatsby-plugin-page-creator@2.0.5) (2018-11-29)
+
+**Note:** Version bump only for package gatsby-plugin-page-creator
+
+<a name="2.0.4"></a>
+
+## [2.0.4](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-page-creator@2.0.3...gatsby-plugin-page-creator@2.0.4) (2018-11-08)
+
+**Note:** Version bump only for package gatsby-plugin-page-creator
+
 <a name="2.0.3"></a>
 
 ## [2.0.3](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-page-creator@2.0.2...gatsby-plugin-page-creator@2.0.3) (2018-11-06)